requirements.txt
annoy==1.17.3
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
scikit_learn==1.3.0
streamlit==1.25.0
tensorflow==2.13.0
tqdm==4.65.0
